requirements.txt
pymupdf>=1.23.0
PyPDF2>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
